refactor(desktop): extract version comparison into utils/version
The yt-dlp updater and the incoming v2 handover bridge both need dotted-numeric
version comparison. Leaving it in ytdlp-spawn would make the bridge import the
downloader's whole module graph behind a fifteen-line helper, so the two
functions move to a dependency-free util and ytdlp-spawn re-exports them.
